backend/foodgram/requirements.txt
django==2.2.16
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
django-filter==21.1
gunicorn==20.0.4
pillow==8.2.0
python-dotenv==0.19.2
psycopg2-binary==2.8.6